Only treat space, `\t`, `\n`, and `\r` as whitespace
The `\s` character class in RegExp includes several other characters
such as `\v`, `\f`, or various space variants. However, in XML only 4
characters are considered whitespace: ` ` (space), `\t`, `\n`, `\r`.

This commit adjusts the pattern so that only these four are matched. It
also adds whitespace characters to the formatting fixture to verify that
they are formatted properly. The first two paragraphs contain tab, space,
and newline characters. The carriage return character was omitted, since
it's quite rare and most text editors convert it to `\n` when saving a
file, so it'd likely be replaced once the fixture is edited again. The
third paragraph contains all other whitespace characters that the `\s`
character class includes.